Network.AWS.CognitoIdentity

Description

Amazon Cognito is a web service that facilitates the delivery of scoped, temporary credentials to mobile devices or other untrusted environments. Amazon Cognito uniquely identifies a device or user and supplies the user with a consistent identity throughout the lifetime of an application. Amazon Cognito lets users authenticate with third-party identity providers (Facebook, Google, or Login with Amazon). As a developer, you decide which identity providers to trust. You can also choose to support unauthenticated access from your application. Your users are provided with Cognito tokens that uniquely identify their device and any information provided about third-party logins.

Documentation

module Network.AWS.CognitoIdentity.CreateIdentityPool

module Network.AWS.CognitoIdentity.DeleteIdentityPool

module Network.AWS.CognitoIdentity.DescribeIdentity

module Network.AWS.CognitoIdentity.DescribeIdentityPool

module Network.AWS.CognitoIdentity.GetCredentialsForIdentity

module Network.AWS.CognitoIdentity.GetId

module Network.AWS.CognitoIdentity.GetIdentityPoolRoles

module Network.AWS.CognitoIdentity.GetOpenIdToken

module Network.AWS.CognitoIdentity.GetOpenIdTokenForDeveloperIdentity

module Network.AWS.CognitoIdentity.ListIdentities

module Network.AWS.CognitoIdentity.ListIdentityPools

module Network.AWS.CognitoIdentity.LookupDeveloperIdentity

module Network.AWS.CognitoIdentity.MergeDeveloperIdentities

module Network.AWS.CognitoIdentity.SetIdentityPoolRoles

module Network.AWS.CognitoIdentity.Types

module Network.AWS.CognitoIdentity.UnlinkDeveloperIdentity

module Network.AWS.CognitoIdentity.UnlinkIdentity

module Network.AWS.CognitoIdentity.UpdateIdentityPool