Network.AWS.CognitoIdentity.CreateIdentityPool

Description

Creates a new identity pool. The identity pool is a store of user identity information that is specific to your AWS account. The limit on identity pools is 60 per account.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_CreateIdentityPool.html

Request

data CreateIdentityPool

Request constructor

createIdentityPool

Arguments

:: Text
cipIdentityPoolName

-> Bool
cipAllowUnauthenticatedIdentities

-> CreateIdentityPool

CreateIdentityPool constructor.

The fields accessible through corresponding lenses are:

cipAllowUnauthenticatedIdentities :: Bool

cipDeveloperProviderName :: Maybe Text

cipIdentityPoolName :: Text

cipOpenIdConnectProviderARNs :: [Text]

cipSupportedLoginProviders :: HashMap Text Text

Request lenses

cipAllowUnauthenticatedIdentities :: Lens' CreateIdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

cipDeveloperProviderName :: Lens' CreateIdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore ('_'), and dash (-).

Once you have set a developer provider name, you cannot change it. Please take care in setting this parameter.

cipIdentityPoolName :: Lens' CreateIdentityPool Text

A string that you provide.

cipOpenIdConnectProviderARNs :: Lens' CreateIdentityPool [Text]

A list of OpendID Connect provider ARNs.

cipSupportedLoginProviders :: Lens' CreateIdentityPool (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

Response

data CreateIdentityPoolResponse

Response constructor

createIdentityPoolResponse

Arguments

:: Text
ciprIdentityPoolId

-> Text
ciprIdentityPoolName

-> Bool
ciprAllowUnauthenticatedIdentities

-> CreateIdentityPoolResponse

CreateIdentityPoolResponse constructor.

The fields accessible through corresponding lenses are:

ciprAllowUnauthenticatedIdentities :: Bool

ciprDeveloperProviderName :: Maybe Text

ciprIdentityPoolId :: Text

ciprIdentityPoolName :: Text

ciprOpenIdConnectProviderARNs :: [Text]

ciprSupportedLoginProviders :: HashMap Text Text

Response lenses

ciprAllowUnauthenticatedIdentities :: Lens' CreateIdentityPoolResponse Bool

TRUE if the identity pool supports unauthenticated logins.

ciprDeveloperProviderName :: Lens' CreateIdentityPoolResponse (Maybe Text)

The "domain" by which Cognito will refer to your users.

ciprIdentityPoolId :: Lens' CreateIdentityPoolResponse Text

An identity pool ID in the format REGION:GUID.

ciprIdentityPoolName :: Lens' CreateIdentityPoolResponse Text

A string that you provide.

ciprOpenIdConnectProviderARNs :: Lens' CreateIdentityPoolResponse [Text]

A list of OpendID Connect provider ARNs.

ciprSupportedLoginProviders :: Lens' CreateIdentityPoolResponse (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.