Network.AWS.CognitoIdentity.DeleteIdentities

Description

Deletes identities from an identity pool. You can specify a list of 1-60 identities that you want to delete.

You must use AWS Developer credentials to call this API.

Creating a Request

deleteIdentities

Arguments

:: NonEmpty Text
diIdentityIdsToDelete

-> DeleteIdentities

Creates a value of DeleteIdentities with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diIdentityIdsToDelete - A list of 1-60 identities that you want to delete.

data DeleteIdentities

Input to the DeleteIdentities action.

See: deleteIdentities smart constructor.

Request Lenses

diIdentityIdsToDelete :: Lens' DeleteIdentities (NonEmpty Text)

A list of 1-60 identities that you want to delete.

Destructuring the Response

deleteIdentitiesResponse

Arguments

:: Int
dirsResponseStatus

-> DeleteIdentitiesResponse

Creates a value of DeleteIdentitiesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dirsUnprocessedIdentityIds - An array of UnprocessedIdentityId objects, each of which contains an ErrorCode and IdentityId.

dirsResponseStatus - -- | The response status code.

data DeleteIdentitiesResponse

Returned in response to a successful DeleteIdentities operation.

See: deleteIdentitiesResponse smart constructor.

Response Lenses

dirsUnprocessedIdentityIds :: Lens' DeleteIdentitiesResponse [UnprocessedIdentityId]

An array of UnprocessedIdentityId objects, each of which contains an ErrorCode and IdentityId.

dirsResponseStatus :: Lens' DeleteIdentitiesResponse Int

- | The response status code.