Network.AWS.CognitoIdentity.DescribeIdentity

Description

Returns metadata related to the given identity, including when the identity was created and any associated linked logins.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_DescribeIdentity.html

Request

data DescribeIdentity

Request constructor

describeIdentity

Arguments

:: Text
diIdentityId

-> DescribeIdentity

DescribeIdentity constructor.

The fields accessible through corresponding lenses are:

diIdentityId :: Text

Request lenses

diIdentityId :: Lens' DescribeIdentity Text

A unique identifier in the format REGION:GUID.

Response

data DescribeIdentityResponse

Response constructor

describeIdentityResponse :: DescribeIdentityResponse

DescribeIdentityResponse constructor.

The fields accessible through corresponding lenses are:

dirCreationDate :: Maybe UTCTime

dirIdentityId :: Maybe Text

dirLastModifiedDate :: Maybe UTCTime

dirLogins :: [Text]

Response lenses

dirCreationDate :: Lens' DescribeIdentityResponse (Maybe UTCTime)

Date on which the identity was created.

dirIdentityId :: Lens' DescribeIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

dirLastModifiedDate :: Lens' DescribeIdentityResponse (Maybe UTCTime)

Date on which the identity was last modified.

dirLogins :: Lens' DescribeIdentityResponse [Text]

A set of optional name-value pairs that map provider names to provider tokens.