Network.AWS.CognitoIdentity.GetCredentialsForIdentity

Description

Returns credentials for the the provided identity ID. Any provided logins will be validated against supported login providers. If the token is for cognito-identity.amazonaws.com, it will be passed through to AWS Security Token Service with the appropriate role for the token.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_GetCredentialsForIdentity.html

Request

data GetCredentialsForIdentity

Request constructor

getCredentialsForIdentity

Arguments

:: Text
gcfiIdentityId

-> GetCredentialsForIdentity

GetCredentialsForIdentity constructor.

The fields accessible through corresponding lenses are:

gcfiIdentityId :: Text

gcfiLogins :: HashMap Text Text

Request lenses

gcfiIdentityId :: Lens' GetCredentialsForIdentity Text

A unique identifier in the format REGION:GUID.

gcfiLogins :: Lens' GetCredentialsForIdentity (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens.

Response

data GetCredentialsForIdentityResponse

Response constructor

getCredentialsForIdentityResponse :: GetCredentialsForIdentityResponse

GetCredentialsForIdentityResponse constructor.

The fields accessible through corresponding lenses are:

gcfirCredentials :: Maybe Credentials

gcfirIdentityId :: Maybe Text

Response lenses

gcfirCredentials :: Lens' GetCredentialsForIdentityResponse (Maybe Credentials)

Credentials for the the provided identity ID.

gcfirIdentityId :: Lens' GetCredentialsForIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.