Network.AWS.CognitoIdentity.GetId

Description

Generates (or retrieves) a Cognito ID. Supplying multiple logins will create an implicit linked account.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_GetId.html

Request

data GetId

Request constructor

getId

Arguments

:: Text
giIdentityPoolId

-> GetId

GetId constructor.

The fields accessible through corresponding lenses are:

giAccountId :: Maybe Text

giIdentityPoolId :: Text

giLogins :: HashMap Text Text

Request lenses

giAccountId :: Lens' GetId (Maybe Text)

A standard AWS account ID (9+ digits).

giIdentityPoolId :: Lens' GetId Text

An identity pool ID in the format REGION:GUID.

giLogins :: Lens' GetId (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens.

The available provider names for Logins are as follows: Facebook: 'graph.facebook.com' Google: 'accounts.google.com' Amazon: 'www.amazon.com'

Response

data GetIdResponse

Response constructor

getIdResponse :: GetIdResponse

GetIdResponse constructor.

The fields accessible through corresponding lenses are:

girIdentityId :: Maybe Text

Response lenses

girIdentityId :: Lens' GetIdResponse (Maybe Text)

A unique identifier in the format REGION:GUID.