Network.AWS.CognitoIdentity.GetIdentityPoolRoles

Description

Gets the roles for an identity pool.

You must use AWS Developer credentials to call this API.

Creating a Request

getIdentityPoolRoles

Arguments

:: Text
giprIdentityPoolId

-> GetIdentityPoolRoles

Creates a value of GetIdentityPoolRoles with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giprIdentityPoolId - An identity pool ID in the format REGION:GUID.

data GetIdentityPoolRoles

Input to the GetIdentityPoolRoles action.

See: getIdentityPoolRoles smart constructor.

Request Lenses

giprIdentityPoolId :: Lens' GetIdentityPoolRoles Text

An identity pool ID in the format REGION:GUID.

Destructuring the Response

getIdentityPoolRolesResponse

Arguments

:: Int
giprrsResponseStatus

-> GetIdentityPoolRolesResponse

Creates a value of GetIdentityPoolRolesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giprrsRoles - The map of roles associated with this pool. Currently only authenticated and unauthenticated roles are supported.

giprrsIdentityPoolId - An identity pool ID in the format REGION:GUID.

giprrsRoleMappings - How users for a specific identity provider are to mapped to roles. This is a String-to-RoleMapping object map. The string identifies the identity provider, for example, "graph.facebook.com" or "cognito-idp-east-1.amazonaws.com/us-east-1_abcdefghi:app_client_id".

giprrsResponseStatus - -- | The response status code.

data GetIdentityPoolRolesResponse

Returned in response to a successful GetIdentityPoolRoles operation.

See: getIdentityPoolRolesResponse smart constructor.

Response Lenses

giprrsRoles :: Lens' GetIdentityPoolRolesResponse (HashMap Text Text)

The map of roles associated with this pool. Currently only authenticated and unauthenticated roles are supported.

giprrsIdentityPoolId :: Lens' GetIdentityPoolRolesResponse (Maybe Text)

An identity pool ID in the format REGION:GUID.

giprrsRoleMappings :: Lens' GetIdentityPoolRolesResponse (HashMap Text RoleMapping)

How users for a specific identity provider are to mapped to roles. This is a String-to-RoleMapping object map. The string identifies the identity provider, for example, "graph.facebook.com" or "cognito-idp-east-1.amazonaws.com/us-east-1_abcdefghi:app_client_id".

giprrsResponseStatus :: Lens' GetIdentityPoolRolesResponse Int

- | The response status code.