Network.AWS.CognitoIdentity.GetOpenIdTokenForDeveloperIdentity

Description

Registers (or retrieves) a Cognito IdentityId and an OpenID Connect token for a user authenticated by your backend authentication process. Supplying multiple logins will create an implicit linked account. You can only specify one developer provider as part of the Logins map, which is linked to the identity pool. The developer provider is the "domain" by which Cognito will refer to your users.

You can use GetOpenIdTokenForDeveloperIdentity to create a new identity and to link new logins (that is, user credentials issued by a public provider or developer provider) to an existing identity. When you want to create a new identity, the IdentityId should be null. When you want to associate a new login with an existing authenticated/unauthenticated identity, you can do so by providing the existing IdentityId. This API will create the identity in the specified IdentityPoolId.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_GetOpenIdTokenForDeveloperIdentity.html

Request

data GetOpenIdTokenForDeveloperIdentity

Request constructor

getOpenIdTokenForDeveloperIdentity

Arguments

:: Text
goitfdiIdentityPoolId

-> GetOpenIdTokenForDeveloperIdentity

GetOpenIdTokenForDeveloperIdentity constructor.

The fields accessible through corresponding lenses are:

goitfdiIdentityId :: Maybe Text

goitfdiIdentityPoolId :: Text

goitfdiLogins :: HashMap Text Text

goitfdiTokenDuration :: Maybe Natural

Request lenses

goitfdiIdentityId :: Lens' GetOpenIdTokenForDeveloperIdentity (Maybe Text)

A unique identifier in the format REGION:GUID.

goitfdiIdentityPoolId :: Lens' GetOpenIdTokenForDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

goitfdiLogins :: Lens' GetOpenIdTokenForDeveloperIdentity (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens. Each name-value pair represents a user from a public provider or developer provider. If the user is from a developer provider, the name-value pair will follow the syntax '"developer_provider_name":"developer_user_identifier"'. The developer provider is the "domain" by which Cognito will refer to your users; you provided this domain while creating/updating the identity pool. The developer user identifier is an identifier from your backend that uniquely identifies a user. When you create an identity pool, you can specify the supported logins.

goitfdiTokenDuration :: Lens' GetOpenIdTokenForDeveloperIdentity (Maybe Natural)

The expiration time of the token, in seconds. You can specify a custom expiration time for the token so that you can cache it. If you don't provide an expiration time, the token is valid for 15 minutes. You can exchange the token with Amazon STS for temporary AWS credentials, which are valid for a maximum of one hour. The maximum token duration you can set is 24 hours. You should take care in setting the expiration time for a token, as there are significant security implications: an attacker could use a leaked token to access your AWS resources for the token's duration.

Response

data GetOpenIdTokenForDeveloperIdentityResponse

Response constructor

getOpenIdTokenForDeveloperIdentityResponse :: GetOpenIdTokenForDeveloperIdentityResponse

GetOpenIdTokenForDeveloperIdentityResponse constructor.

The fields accessible through corresponding lenses are:

goitfdirIdentityId :: Maybe Text

goitfdirToken :: Maybe Text

Response lenses

goitfdirIdentityId :: Lens' GetOpenIdTokenForDeveloperIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

goitfdirToken :: Lens' GetOpenIdTokenForDeveloperIdentityResponse (Maybe Text)

An OpenID token.