Network.AWS.CognitoIdentity.ListIdentityPools

Description

Lists all of the Cognito identity pools registered for your account. ListIdentityPools The following example shows a request and a response for a ListIdentityPools operation. { MaxResults: 10 } { IdentityPools: [ { IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1", IdentityPoolName: MyPool }, { IdentityPoolId: "us-east-1:f212b602-a526-4557-af13-8eedEXAMPLE2", IdentityPoolName: MyPool2 } ] }.

Request

data ListIdentityPools

Input to the ListIdentityPools action.

Request constructor

listIdentityPools

Arguments

:: Integer
lipMaxResults

-> ListIdentityPools

Smart constructor for the minimum required parameters to construct a valid ListIdentityPools request.

The fields accessible through corresponding lenses are:

MaxResults :: Integer

NextToken :: Maybe Text

Request lenses

lipMaxResults :: Lens' ListIdentityPools Integer

The maximum number of identities to return.

lipNextToken :: Lens' ListIdentityPools (Maybe Text)

A pagination token.

Response

data ListIdentityPoolsResponse

The result of a successful ListIdentityPools action.

Response constructor

listIdentityPoolsResponse :: ListIdentityPoolsResponse

Smart constructor for the minimum required parameters to construct a valid ListIdentityPoolsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IdentityPools :: [IdentityPoolShortDescription]

NextToken :: Maybe Text

Response lenses

liprIdentityPools :: Lens' ListIdentityPoolsResponse [IdentityPoolShortDescription]

The identity pools returned by the ListIdentityPools action.

liprNextToken :: Lens' ListIdentityPoolsResponse (Maybe Text)

A pagination token.