Network.AWS.CognitoIdentity.LookupDeveloperIdentity

Description

Retrieves the IdentityID associated with a DeveloperUserIdentifier or the list of DeveloperUserIdentifiers associated with an IdentityId for an existing identity. Either IdentityID or DeveloperUserIdentifier must not be null. If you supply only one of these values, the other value will be searched in the database and returned as a part of the response. If you supply both, DeveloperUserIdentifier will be matched against IdentityID. If the values are verified against the database, the response returns both values and is the same as the request. Otherwise a ResourceConflictException is thrown.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_LookupDeveloperIdentity.html

Request

data LookupDeveloperIdentity

Request constructor

lookupDeveloperIdentity

Arguments

:: Text
ldiIdentityPoolId

-> LookupDeveloperIdentity

LookupDeveloperIdentity constructor.

The fields accessible through corresponding lenses are:

ldiDeveloperUserIdentifier :: Maybe Text

ldiIdentityId :: Maybe Text

ldiIdentityPoolId :: Text

ldiMaxResults :: Maybe Natural

ldiNextToken :: Maybe Text

Request lenses

ldiDeveloperUserIdentifier :: Lens' LookupDeveloperIdentity (Maybe Text)

A unique ID used by your backend authentication process to identify a user. Typically, a developer identity provider would issue many developer user identifiers, in keeping with the number of users.

ldiIdentityId :: Lens' LookupDeveloperIdentity (Maybe Text)

A unique identifier in the format REGION:GUID.

ldiIdentityPoolId :: Lens' LookupDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

ldiMaxResults :: Lens' LookupDeveloperIdentity (Maybe Natural)

The maximum number of identities to return.

ldiNextToken :: Lens' LookupDeveloperIdentity (Maybe Text)

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

Response

data LookupDeveloperIdentityResponse

Response constructor

lookupDeveloperIdentityResponse :: LookupDeveloperIdentityResponse

LookupDeveloperIdentityResponse constructor.

The fields accessible through corresponding lenses are:

ldirDeveloperUserIdentifierList :: [Text]

ldirIdentityId :: Maybe Text

ldirNextToken :: Maybe Text

Response lenses

ldirDeveloperUserIdentifierList :: Lens' LookupDeveloperIdentityResponse [Text]

This is the list of developer user identifiers associated with an identity ID. Cognito supports the association of multiple developer user identifiers with an identity ID.

ldirIdentityId :: Lens' LookupDeveloperIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

ldirNextToken :: Lens' LookupDeveloperIdentityResponse (Maybe Text)

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.