Network.AWS.CognitoIdentity.MergeDeveloperIdentities

Description

Merges two users having different IdentityIds, existing in the same identity pool, and identified by the same developer provider. You can use this action to request that discrete users be merged and identified as a single user in the Cognito environment. Cognito associates the given source user (SourceUserIdentifier) with the IdentityId of the DestinationUserIdentifier. Only developer-authenticated users can be merged. If the users to be merged are associated with the same public provider, but as two different users, an exception will be thrown.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_MergeDeveloperIdentities.html

Request

data MergeDeveloperIdentities

Request constructor

mergeDeveloperIdentities

Arguments

:: Text
mdiSourceUserIdentifier

-> Text
mdiDestinationUserIdentifier

-> Text
mdiDeveloperProviderName

-> Text
mdiIdentityPoolId

-> MergeDeveloperIdentities

MergeDeveloperIdentities constructor.

The fields accessible through corresponding lenses are:

mdiDestinationUserIdentifier :: Text

mdiDeveloperProviderName :: Text

mdiIdentityPoolId :: Text

mdiSourceUserIdentifier :: Text

Request lenses

mdiDestinationUserIdentifier :: Lens' MergeDeveloperIdentities Text

User identifier for the destination user. The value should be a DeveloperUserIdentifier.

mdiDeveloperProviderName :: Lens' MergeDeveloperIdentities Text

The "domain" by which Cognito will refer to your users. This is a (pseudo) domain name that you provide while creating an identity pool. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

mdiIdentityPoolId :: Lens' MergeDeveloperIdentities Text

An identity pool ID in the format REGION:GUID.

mdiSourceUserIdentifier :: Lens' MergeDeveloperIdentities Text

User identifier for the source user. The value should be a DeveloperUserIdentifier.

Response

data MergeDeveloperIdentitiesResponse

Response constructor

mergeDeveloperIdentitiesResponse :: MergeDeveloperIdentitiesResponse

MergeDeveloperIdentitiesResponse constructor.

The fields accessible through corresponding lenses are:

mdirIdentityId :: Maybe Text

Response lenses

mdirIdentityId :: Lens' MergeDeveloperIdentitiesResponse (Maybe Text)

A unique identifier in the format REGION:GUID.