Network.AWS.CognitoIdentity.Monadic

Description

This module is provided for convenience. It offers an alternative to the common idiom of supplying required fields to an operations's smart constructor, using the operation's lenses to modify additional fields, and then sending or paginating the request.

As an example: using Network.AWS.CognitoIdentity with the smart constructor and basic lens syntax, before explicitly calling send:

import Control.Monad.Trans.AWS
import Network.AWS.CognitoIdentity

send $ (mkOperationName w x)
     & onLensField1 .~ y
     & onLensField2 .~ z

Versus using Network.AWS.CognitoIdentity.Monadic with the State operator variants from Control.Lens.Setter such as .= to modify any additional request parameters before sending:

import Control.Applicative
import Network.AWS.CognitoIdentity.Monadic

operationName w x $ do
    onLensField1 .= y
    onLensField2 .= z

-- Or to void any additional parameters outside of those required using empty:
operationName w x empty

CreateIdentityPool

Creates a new identity pool. The identity pool is a store of user identity information that is specific to your AWS account. CreateIdentityPool The following example shows a request and response for a CreateIdentityPool operation. { IdentityPoolName: MyIdentityPool, IdentityPoolDescription: "My identity pool", Unauthenticated: true, SupportedLoginProviders: { "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID, "www.amazon.com": Amazon_App_ID } } { IdentityPoolDescription: "My identity pool", IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyIdentityPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, Unauthenticated: true }.

See: CreateIdentityPool

createIdentityPool

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
cipIdentityPoolName

-> Bool
cipAllowUnauthenticatedIdentities

-> State CreateIdentityPool a

-> m CreateIdentityPoolResponse

createIdentityPoolCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
cipIdentityPoolName

-> Bool
cipAllowUnauthenticatedIdentities

-> State CreateIdentityPool a

-> m (Either ServiceEr CreateIdentityPoolResponse)

DeleteIdentityPool

Deletes a user pool. Once a pool is deleted, users will not be able to authenticate with the pool. DeleteIdentityPool The following is an example of a DeleteIdentityPool request. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE" }.

See: DeleteIdentityPool

deleteIdentityPool

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dipIdentityPoolId

-> m DeleteIdentityPoolResponse

deleteIdentityPoolCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dipIdentityPoolId

-> m (Either ServiceEr DeleteIdentityPoolResponse)

DescribeIdentityPool

Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users. DescribeIdentityPool The following are an example request and response for the DescribeIdentityPool operation. { IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1" } { IdentityPoolDescription: "My identity pool", IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1", IdentityPoolName: MyIdentityPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, Unauthenticated: true }.

See: DescribeIdentityPool

describeIdentityPool

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dip1IdentityPoolId

-> m DescribeIdentityPoolResponse

describeIdentityPoolCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dip1IdentityPoolId

-> m (Either ServiceEr DescribeIdentityPoolResponse)

GetId

Generates (or retrieves) a Cognito ID. Supplying multiple logins will create an implicit linked account. GetId The following example shows a GetId request for an unauthenticated identity. { AccountId: "123456789012;", IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1" } { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3" }.

See: GetId

getId

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
giAccountId

-> Text
giIdentityPoolId

-> State GetId a

-> m GetIdResponse

getIdCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
giAccountId

-> Text
giIdentityPoolId

-> State GetId a

-> m (Either ServiceEr GetIdResponse)

GetOpenIdToken

Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned from GetId. You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link. GetOpenIdToken The following examples show a GetOpenIdToken request and response, without the optional login values. { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3" } { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3", Token: "EXAmpLeTOkENUzUxMiIsInR5cCI6IkpXUyIsImtpZCI6InVzLWVhc3QtMTEifQ.eyJleHAiOjE0MDI2Njg0NTAsInN1YiI6InVzLWVhc3QtMTo5MjFhMzg0My0yZGQ2LTQ2YjgtYWIyZi1jNjc5NTUyZTM3MWUiLCJhdWQiOiJ1cy1lYXN0LTE6YWY0MzExY2EtODM1ZS00YjQ5LTgxNGMtMjI5MGQ4ZDU1YTZmIiwiaXNzIjoiaHR0cHM6Ly9jb2duaXRvLWlkZW50aXR5LXB1YmxpYy1pYWQtYmV0YS5hbWF6b24uY29tIiwiaWF0IjoxNDAyNjY3ODUwLCJhbXIiOlsidW5hdXRoZW50aWNhdGVkIl19.faWdRGsKxT8YqTBnAow1fNyXE57kjScKQ0lyFpFAUIl6VNVV-nQ_QD8XKHB_pAY2UNtxYFDoGhHrL3cqH_FLSfRLG-X3EaIrCsr0A6KIW7X69wsAxJQB-EvYru0UhDpcPaDyQUXTHFmRP9bzJMsSLi7nXm-OD4DCujX3vKwOhlSymbH9KbAG105t3_G_a8tsUbCV488nvlrA-7Omp0D18T1__XeZttldW1GODOK2OY2bK5-3eyodcqbVXaPTotxO5PTlYpzuMS1XfTejC8LJ2DocP_eBT7xhSr2qkro9Y6uCDph_-6ttYrXRaaLKZv3v1Lz6PGHrsPhJdK_bYRHhdg" }.

See: GetOpenIdToken

getOpenIdToken

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
goitIdentityId

-> State GetOpenIdToken a

-> m GetOpenIdTokenResponse

getOpenIdTokenCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
goitIdentityId

-> State GetOpenIdToken a

-> m (Either ServiceEr GetOpenIdTokenResponse)

ListIdentities

Lists the identities in a pool. ListIdentities The following are examples of a request and a response for the ListIdentities action. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", MaxResults: 10 } { Identities: [ { IdentityId: "us-east-1:2345a6b7-8cc3-4a60-8aeb-e11bEXAMPLE4" }, { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3" }, { IdentityId: "us-east-1:921a3843-2dd6-46b8-ab2f-c679EXAMPLE5" } ], IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE" }.

See: ListIdentities

listIdentities

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
liIdentityPoolId

-> Integer
liMaxResults

-> State ListIdentities a

-> m ListIdentitiesResponse

listIdentitiesCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
liIdentityPoolId

-> Integer
liMaxResults

-> State ListIdentities a

-> m (Either ServiceEr ListIdentitiesResponse)

ListIdentityPools

Lists all of the Cognito identity pools registered for your account. ListIdentityPools The following example shows a request and a response for a ListIdentityPools operation. { MaxResults: 10 } { IdentityPools: [ { IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1", IdentityPoolName: MyPool }, { IdentityPoolId: "us-east-1:f212b602-a526-4557-af13-8eedEXAMPLE2", IdentityPoolName: MyPool2 } ] }.

See: ListIdentityPools

listIdentityPools

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Integer
lipMaxResults

-> State ListIdentityPools a

-> m ListIdentityPoolsResponse

listIdentityPoolsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Integer
lipMaxResults

-> State ListIdentityPools a

-> m (Either ServiceEr ListIdentityPoolsResponse)

UnlinkIdentity

Unlinks a federated identity from an existing account. Unlinked logins will be considered new identities next time they are seen. Removing the last linked login will make this identity inaccessible.

See: UnlinkIdentity

unlinkIdentity

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uiIdentityId

-> Map Text Text
uiLogins

-> [Text]
uiLoginsToRemove

-> m UnlinkIdentityResponse

unlinkIdentityCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uiIdentityId

-> Map Text Text
uiLogins

-> [Text]
uiLoginsToRemove

-> m (Either ServiceEr UnlinkIdentityResponse)

UpdateIdentityPool

Updates a user pool. UpdateIdentityPool The following are a request and response for the UpdateIdentityPool action. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyUpdatedPool, IdentityPoolDescription: "An identity pool that needs updating", Unauthenticated: true, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID } } { IdentityPoolDescription: "An identity pool that needs updating", IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyUpdatedPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, AllowUnauthenticatedIdentities: true }.

See: UpdateIdentityPool

updateIdentityPool

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
uipIdentityPoolId

-> Text
uipIdentityPoolName

-> Bool
uipAllowUnauthenticatedIdentities

-> State UpdateIdentityPool a

-> m UpdateIdentityPoolResponse

updateIdentityPoolCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
uipIdentityPoolId

-> Text
uipIdentityPoolName

-> Bool
uipAllowUnauthenticatedIdentities

-> State UpdateIdentityPool a

-> m (Either ServiceEr UpdateIdentityPoolResponse)

Re-exported

module Network.AWS.CognitoIdentity

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()