Network.AWS.CognitoIdentity.Types

Description

Amazon Cognito is a web service that facilitates the delivery of scoped, temporary credentials to mobile devices or other untrusted environments. Amazon Cognito uniquely identifies a device or user and supplies the user with a consistent identity throughout the lifetime of an application. Amazon Cognito lets users authenticate with third-party identity providers (Facebook, Google, or Login with Amazon). As a developer, you decide which identity providers to trust. You can also choose to support unauthenticated access from your application. Your users are provided with Cognito tokens that uniquely identify their device and any information provided about third-party logins.

Service

data CognitoIdentity

Supported version (2014-06-30) of the Amazon Cognito Identity service.

Errors

data CognitoIdentityError

A sum type representing possible errors returned by the CognitoIdentity service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

CognitoIdentityClient HttpException

CognitoIdentitySerializer String

CognitoIdentityService String

InternalErrorException
Thrown when the service encounters an error during processing the request.

Fields

_ieeMessage :: Maybe Text

InvalidParameterException
Thrown for missing or bad input parameter(s).

Fields

_ipeMessage :: Maybe Text

LimitExceededException
Thrown when the total number of user pools has exceeded a preset limit.

Fields

_leeMessage :: Maybe Text

NotAuthorizedException
Thrown when a user is not authorized to access the requested resource.

Fields

_naeMessage :: Maybe Text

ResourceConflictException
Thrown when a user tries to use a login which is already linked to another account.

Fields

_rceMessage :: Maybe Text

ResourceNotFoundException
Thrown when the requested resource (for example, a dataset or record) does not exist.

Fields

_rnfeMessage :: Maybe Text

TooManyRequestsException
Thrown when a request is throttled.

Fields

_tmreMessage :: Maybe Text

_CognitoIdentityClient :: Prism' CognitoIdentityError HttpException

See: CognitoIdentityClient

_CognitoIdentitySerializer :: Prism' CognitoIdentityError String

See: CognitoIdentitySerializer

_CognitoIdentityService :: Prism' CognitoIdentityError String

See: CognitoIdentityService

_InternalErrorException :: Prism' CognitoIdentityError (Maybe Text)

Thrown when the service encounters an error during processing the request.

See: InternalErrorException

_InvalidParameterException :: Prism' CognitoIdentityError (Maybe Text)

Thrown for missing or bad input parameter(s).

See: InvalidParameterException

_LimitExceededException :: Prism' CognitoIdentityError (Maybe Text)

Thrown when the total number of user pools has exceeded a preset limit.

See: LimitExceededException

_NotAuthorizedException :: Prism' CognitoIdentityError (Maybe Text)

Thrown when a user is not authorized to access the requested resource.

See: NotAuthorizedException

_ResourceConflictException :: Prism' CognitoIdentityError (Maybe Text)

Thrown when a user tries to use a login which is already linked to another account.

See: ResourceConflictException

_ResourceNotFoundException :: Prism' CognitoIdentityError (Maybe Text)

Thrown when the requested resource (for example, a dataset or record) does not exist.

See: ResourceNotFoundException

_TooManyRequestsException :: Prism' CognitoIdentityError (Maybe Text)

Thrown when a request is throttled.

See: TooManyRequestsException

IdentityDescription

data IdentityDescription

A description of the identity.

identityDescription :: IdentityDescription

Smart constructor for the minimum required fields to construct a valid IdentityDescription data type.

IdentityDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IdentityId :: Maybe Text

Logins :: [Text]

idIdentityId :: Lens' IdentityDescription (Maybe Text)

A unique identifier in the format REGION:GUID.

idLogins :: Lens' IdentityDescription [Text]

A set of optional name/value pairs that map provider names to provider tokens.

IdentityPoolShortDescription

data IdentityPoolShortDescription

A description of the identity pool.

identityPoolShortDescription :: IdentityPoolShortDescription

Smart constructor for the minimum required fields to construct a valid IdentityPoolShortDescription data type.

IdentityPoolShortDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IdentityPoolId :: Maybe Text

IdentityPoolName :: Maybe Text

ipsdIdentityPoolId :: Lens' IdentityPoolShortDescription (Maybe Text)

An identity pool ID in the format REGION:GUID.

ipsdIdentityPoolName :: Lens' IdentityPoolShortDescription (Maybe Text)

A string that you provide.