Network.AWS.CognitoIdentity.UnlinkDeveloperIdentity

Description

Unlinks a DeveloperUserIdentifier from an existing identity. Unlinked developer users will be considered new identities next time they are seen. If, for a given Cognito identity, you remove all federated identities as well as the developer user identifier, the Cognito identity becomes inaccessible.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_UnlinkDeveloperIdentity.html

Request

data UnlinkDeveloperIdentity

Request constructor

unlinkDeveloperIdentity

Arguments

:: Text
udiIdentityId

-> Text
udiIdentityPoolId

-> Text
udiDeveloperProviderName

-> Text
udiDeveloperUserIdentifier

-> UnlinkDeveloperIdentity

UnlinkDeveloperIdentity constructor.

The fields accessible through corresponding lenses are:

udiDeveloperProviderName :: Text

udiDeveloperUserIdentifier :: Text

udiIdentityId :: Text

udiIdentityPoolId :: Text

Request lenses

udiDeveloperProviderName :: Lens' UnlinkDeveloperIdentity Text

The "domain" by which Cognito will refer to your users.

udiDeveloperUserIdentifier :: Lens' UnlinkDeveloperIdentity Text

A unique ID used by your backend authentication process to identify a user.

udiIdentityId :: Lens' UnlinkDeveloperIdentity Text

A unique identifier in the format REGION:GUID.

udiIdentityPoolId :: Lens' UnlinkDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

Response

data UnlinkDeveloperIdentityResponse

Response constructor

unlinkDeveloperIdentityResponse :: UnlinkDeveloperIdentityResponse

UnlinkDeveloperIdentityResponse constructor.