Network.AWS.CognitoIdentity.UnlinkIdentity

Description

Unlinks a federated identity from an existing account. Unlinked logins will be considered new identities next time they are seen. Removing the last linked login will make this identity inaccessible.

This is a public API. You do not need any credentials to call this API.

Creating a Request

unlinkIdentity

Arguments

:: Text
uiIdentityId

-> UnlinkIdentity

Creates a value of UnlinkIdentity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uiIdentityId - A unique identifier in the format REGION:GUID.

uiLogins - A set of optional name-value pairs that map provider names to provider tokens.

uiLoginsToRemove - Provider names to unlink from this identity.

data UnlinkIdentity

Input to the UnlinkIdentity action.

See: unlinkIdentity smart constructor.

Request Lenses

uiIdentityId :: Lens' UnlinkIdentity Text

A unique identifier in the format REGION:GUID.

uiLogins :: Lens' UnlinkIdentity (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens.

uiLoginsToRemove :: Lens' UnlinkIdentity [Text]

Provider names to unlink from this identity.

Destructuring the Response

unlinkIdentityResponse :: UnlinkIdentityResponse

Creates a value of UnlinkIdentityResponse with the minimum fields required to make a request.

data UnlinkIdentityResponse

See: unlinkIdentityResponse smart constructor.