Network.AWS.CognitoIdentity.UpdateIdentityPool

Description

Updates a user pool.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_UpdateIdentityPool.html

Request

data UpdateIdentityPool

Request constructor

updateIdentityPool

Arguments

:: Text
uipIdentityPoolId

-> Text
uipIdentityPoolName

-> Bool
uipAllowUnauthenticatedIdentities

-> UpdateIdentityPool

UpdateIdentityPool constructor.

The fields accessible through corresponding lenses are:

uipAllowUnauthenticatedIdentities :: Bool

uipDeveloperProviderName :: Maybe Text

uipIdentityPoolId :: Text

uipIdentityPoolName :: Text

uipOpenIdConnectProviderARNs :: [Text]

uipSupportedLoginProviders :: HashMap Text Text

Request lenses

uipAllowUnauthenticatedIdentities :: Lens' UpdateIdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

uipDeveloperProviderName :: Lens' UpdateIdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

uipIdentityPoolId :: Lens' UpdateIdentityPool Text

An identity pool ID in the format REGION:GUID.

uipIdentityPoolName :: Lens' UpdateIdentityPool Text

A string that you provide.

uipOpenIdConnectProviderARNs :: Lens' UpdateIdentityPool [Text]

A list of OpendID Connect provider ARNs.

uipSupportedLoginProviders :: Lens' UpdateIdentityPool (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

Response

data UpdateIdentityPoolResponse

Response constructor

updateIdentityPoolResponse

Arguments

:: Text
uiprIdentityPoolId

-> Text
uiprIdentityPoolName

-> Bool
uiprAllowUnauthenticatedIdentities

-> UpdateIdentityPoolResponse

UpdateIdentityPoolResponse constructor.

The fields accessible through corresponding lenses are:

uiprAllowUnauthenticatedIdentities :: Bool

uiprDeveloperProviderName :: Maybe Text

uiprIdentityPoolId :: Text

uiprIdentityPoolName :: Text

uiprOpenIdConnectProviderARNs :: [Text]

uiprSupportedLoginProviders :: HashMap Text Text

Response lenses

uiprAllowUnauthenticatedIdentities :: Lens' UpdateIdentityPoolResponse Bool

TRUE if the identity pool supports unauthenticated logins.

uiprDeveloperProviderName :: Lens' UpdateIdentityPoolResponse (Maybe Text)

The "domain" by which Cognito will refer to your users.

uiprIdentityPoolId :: Lens' UpdateIdentityPoolResponse Text

An identity pool ID in the format REGION:GUID.

uiprIdentityPoolName :: Lens' UpdateIdentityPoolResponse Text

A string that you provide.

uiprOpenIdConnectProviderARNs :: Lens' UpdateIdentityPoolResponse [Text]

A list of OpendID Connect provider ARNs.

uiprSupportedLoginProviders :: Lens' UpdateIdentityPoolResponse (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.