Network.AWS.CognitoIdentity.V2014_06_30.CreateIdentityPool

Description

Creates a new identity pool. The identity pool is a store of user identity information that is specific to your AWS account. CreateIdentityPool The following example shows a request and response for a CreateIdentityPool operation. { IdentityPoolName: MyIdentityPool, IdentityPoolDescription: "My identity pool", Unauthenticated: true, SupportedLoginProviders: { "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID, "www.amazon.com": Amazon_App_ID } } { IdentityPoolDescription: "My identity pool", IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyIdentityPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, Unauthenticated: true }.

Documentation

createIdentityPool

Arguments

:: Text
_cipiIdentityPoolName

-> Bool
_cipiAllowUnauthenticatedIdentities

-> CreateIdentityPool

Minimum specification for a CreateIdentityPool request.

data CreateIdentityPool

Constructors

CreateIdentityPool

Fields

_cipiIdentityPoolName :: Text

A string that you provide.

_cipiAllowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

_cipiSupportedLoginProviders :: Map Text Text

Optional key:value pairs mapping provider names to provider app IDs.

cipiSupportedLoginProviders :: Lens' CreateIdentityPool (Map Text Text)

cipiIdentityPoolName :: Lens' CreateIdentityPool Text

cipiAllowUnauthenticatedIdentities :: Lens' CreateIdentityPool Bool

data CreateIdentityPoolResponse

Constructors

CreateIdentityPoolResponse

Fields

_iwIdentityPoolId :: Text

An identity pool ID in the format REGION:GUID.

_iwIdentityPoolName :: Text

A string that you provide.

_iwAllowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

_iwSupportedLoginProviders :: Map Text Text

Optional key:value pairs mapping provider names to provider app IDs.

iwSupportedLoginProviders :: Lens' CreateIdentityPoolResponse (Map Text Text)

iwIdentityPoolName :: Lens' CreateIdentityPoolResponse Text

iwIdentityPoolId :: Lens' CreateIdentityPoolResponse Text

iwAllowUnauthenticatedIdentities :: Lens' CreateIdentityPoolResponse Bool