Network.AWS.CognitoIdentity.V2014_06_30.DeleteIdentityPool

Description

Deletes a user pool. Once a pool is deleted, users will not be able to authenticate with the pool. DeleteIdentityPool The following is an example of a DeleteIdentityPool request. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE" }.

Documentation

data DeleteIdentityPool

Constructors

DeleteIdentityPool

Fields

_dipjIdentityPoolId :: Text

An identity pool ID in the format REGION:GUID.

dipjIdentityPoolId :: Iso' DeleteIdentityPool Text

data DeleteIdentityPoolResponse

Constructors

DeleteIdentityPoolResponse