Network.AWS.CognitoIdentity.V2014_06_30.DescribeIdentityPool

Description

Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users. DescribeIdentityPool The following are an example request and response for the DescribeIdentityPool operation. { IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1" } { IdentityPoolDescription: "My identity pool", IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1", IdentityPoolName: MyIdentityPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, Unauthenticated: true }.

Documentation

data DescribeIdentityPool

Constructors

DescribeIdentityPool

Fields

_dipiIdentityPoolId :: Text

An identity pool ID in the format REGION:GUID.

dipiIdentityPoolId :: Iso' DescribeIdentityPool Text

data DescribeIdentityPoolResponse

Constructors

DescribeIdentityPoolResponse

Fields

_ipIdentityPoolId :: Text

An identity pool ID in the format REGION:GUID.

_ipIdentityPoolName :: Text

A string that you provide.

_ipAllowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

_ipSupportedLoginProviders :: Map Text Text

Optional key:value pairs mapping provider names to provider app IDs.

ipSupportedLoginProviders :: Lens' DescribeIdentityPoolResponse (Map Text Text)

ipIdentityPoolName :: Lens' DescribeIdentityPoolResponse Text

ipIdentityPoolId :: Lens' DescribeIdentityPoolResponse Text

ipAllowUnauthenticatedIdentities :: Lens' DescribeIdentityPoolResponse Bool