Network.AWS.CognitoIdentity.V2014_06_30.GetId

Description

Generates (or retrieves) a Cognito ID. Supplying multiple logins will create an implicit linked account. GetId The following example shows a GetId request for an unauthenticated identity. { AccountId: "123456789012;", IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1" } { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3" }.

Request

data GetId

Input to the GetId action.

Request constructor

mkGetId

Arguments

:: Text
giAccountId

-> Text
giIdentityPoolId

-> GetId

Smart constructor for the minimum required parameters to construct a valid GetId request.

Request lenses

giAccountId :: Lens' GetId Text

A standard AWS account ID (9+ digits).

giIdentityPoolId :: Lens' GetId Text

An identity pool ID in the format REGION:GUID.

giLogins :: Lens' GetId (Map Text Text)

A set of optional name/value pairs that map provider names to provider tokens.

Response

data GetIdResponse

Returned in the response to a GetId request.

Response constructor

mkGetIdResponse :: GetIdResponse

Smart constructor for the minimum required parameters to construct a valid GetIdResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

girIdentityId :: Lens' GetIdResponse (Maybe Text)

A unique identifier in the format REGION:GUID.