Network.AWS.CognitoIdentity.V2014_06_30.GetOpenIdToken

Description

Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned from GetId. You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link. GetOpenIdToken The following examples show a GetOpenIdToken request and response, without the optional login values. { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3" } { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3", Token: "EXAmpLeTOkENUzUxMiIsInR5cCI6IkpXUyIsImtpZCI6InVzLWVhc3QtMTEifQ.eyJleHAiOjE0MDI2Njg0NTAsInN1YiI6InVzLWVhc3QtMTo5MjFhMzg0My0yZGQ2LTQ2YjgtYWIyZi1jNjc5NTUyZTM3MWUiLCJhdWQiOiJ1cy1lYXN0LTE6YWY0MzExY2EtODM1ZS00YjQ5LTgxNGMtMjI5MGQ4ZDU1YTZmIiwiaXNzIjoiaHR0cHM6Ly9jb2duaXRvLWlkZW50aXR5LXB1YmxpYy1pYWQtYmV0YS5hbWF6b24uY29tIiwiaWF0IjoxNDAyNjY3ODUwLCJhbXIiOlsidW5hdXRoZW50aWNhdGVkIl19.faWdRGsKxT8YqTBnAow1fNyXE57kjScKQ0lyFpFAUIl6VNVV-nQ_QD8XKHB_pAY2UNtxYFDoGhHrL3cqH_FLSfRLG-X3EaIrCsr0A6KIW7X69wsAxJQB-EvYru0UhDpcPaDyQUXTHFmRP9bzJMsSLi7nXm-OD4DCujX3vKwOhlSymbH9KbAG105t3_G_a8tsUbCV488nvlrA-7Omp0D18T1__XeZttldW1GODOK2OY2bK5-3eyodcqbVXaPTotxO5PTlYpzuMS1XfTejC8LJ2DocP_eBT7xhSr2qkro9Y6uCDph_-6ttYrXRaaLKZv3v1Lz6PGHrsPhJdK_bYRHhdg" }.

Documentation

getOpenIdToken

Arguments

:: Text
_goitiIdentityId

-> GetOpenIdToken

Minimum specification for a GetOpenIdToken request.

data GetOpenIdToken

Constructors

GetOpenIdToken

Fields

_goitiIdentityId :: Text

A unique identifier in the format REGION:GUID.

_goitiLogins :: Map Text Text

A set of optional name/value pairs that map provider names to provider tokens.

goitiLogins :: Lens' GetOpenIdToken (Map Text Text)

goitiIdentityId :: Lens' GetOpenIdToken Text

data GetOpenIdTokenResponse

Constructors

GetOpenIdTokenResponse

Fields

_goitrIdentityId :: Maybe Text

A unique identifier in the format REGION:GUID. Note that the IdentityId returned may not match the one passed on input.

_goitrToken :: Maybe Text

An OpenID token.

goitrToken :: Lens' GetOpenIdTokenResponse (Maybe Text)

goitrIdentityId :: Lens' GetOpenIdTokenResponse (Maybe Text)