Network.AWS.CognitoIdentity.V2014_06_30.ListIdentities

Description

Lists the identities in a pool. ListIdentities The following are examples of a request and a response for the ListIdentities action. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", MaxResults: 10 } { Identities: [ { IdentityId: "us-east-1:2345a6b7-8cc3-4a60-8aeb-e11bEXAMPLE4" }, { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3" }, { IdentityId: "us-east-1:921a3843-2dd6-46b8-ab2f-c679EXAMPLE5" } ], IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE" }.

Documentation

listIdentities

Arguments

:: Text
_liiIdentityPoolId

-> Integer
_liiMaxResults

-> ListIdentities

Minimum specification for a ListIdentities request.

data ListIdentities

Constructors

ListIdentities

Fields

_liiIdentityPoolId :: Text

An identity pool ID in the format REGION:GUID.

_liiMaxResults :: Integer

The maximum number of identities to return.

_liiNextToken :: Maybe Text

A pagination token.

liiNextToken :: Lens' ListIdentities (Maybe Text)

liiMaxResults :: Lens' ListIdentities Integer

liiIdentityPoolId :: Lens' ListIdentities Text

data ListIdentitiesResponse

Constructors

ListIdentitiesResponse

Fields

_lirIdentities :: [IdentityDescription]

An object containing a set of identities and associated mappings.

_lirIdentityPoolId :: Maybe Text

An identity pool ID in the format REGION:GUID.

_lirNextToken :: Maybe Text

A pagination token.

lirNextToken :: Lens' ListIdentitiesResponse (Maybe Text)

lirIdentityPoolId :: Lens' ListIdentitiesResponse (Maybe Text)

lirIdentities :: Lens' ListIdentitiesResponse [IdentityDescription]