Network.AWS.CognitoIdentity.V2014_06_30.ListIdentityPools

Description

Lists all of the Cognito identity pools registered for your account. ListIdentityPools The following example shows a request and a response for a ListIdentityPools operation. { MaxResults: 10 } { IdentityPools: [ { IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1", IdentityPoolName: MyPool }, { IdentityPoolId: "us-east-1:f212b602-a526-4557-af13-8eedEXAMPLE2", IdentityPoolName: MyPool2 } ] }.

Documentation

listIdentityPools

Arguments

:: Integer
_lipiMaxResults

-> ListIdentityPools

Minimum specification for a ListIdentityPools request.

data ListIdentityPools

Constructors

ListIdentityPools

Fields

_lipiMaxResults :: Integer

The maximum number of identities to return.

_lipiNextToken :: Maybe Text

A pagination token.

lipiNextToken :: Lens' ListIdentityPools (Maybe Text)

lipiMaxResults :: Lens' ListIdentityPools Integer

data ListIdentityPoolsResponse

Constructors

ListIdentityPoolsResponse

Fields

_liprIdentityPools :: [IdentityPoolShortDescription]

The identity pools returned by the ListIdentityPools action.

_liprNextToken :: Maybe Text

A pagination token.

liprNextToken :: Lens' ListIdentityPoolsResponse (Maybe Text)

liprIdentityPools :: Lens' ListIdentityPoolsResponse [IdentityPoolShortDescription]