Network.AWS.CognitoIdentity.V2014_06_30.Types

Description

Amazon Cognito is a web service that facilitates the delivery of scoped, temporary credentials to mobile devices or other untrusted environments. Amazon Cognito uniquely identifies a device or user and supplies the user with a consistent identity throughout the lifetime of an application. Amazon Cognito lets users authenticate with third-party identity providers (Facebook, Google, or Login with Amazon). As a developer, you decide which identity providers to trust. You can also choose to support unauthenticated access from your application. Your users are provided with Cognito tokens that uniquely identify their device and any information provided about third-party logins.

Service

data CognitoIdentity

Supported version (2014-06-30) of the Amazon Cognito Identity service.

IdentityDescription

data IdentityDescription

A description of the identity.

mkIdentityDescription :: IdentityDescription

Smart constructor for the minimum required fields to construct a valid IdentityDescription data type.

IdentityDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IdentityId :: Maybe Text

Logins :: [Text]

idIdentityId :: Lens' IdentityDescription (Maybe Text)

A unique identifier in the format REGION:GUID.

idLogins :: Lens' IdentityDescription [Text]

A set of optional name/value pairs that map provider names to provider tokens.

IdentityPoolShortDescription

data IdentityPoolShortDescription

A description of the identity pool.

mkIdentityPoolShortDescription :: IdentityPoolShortDescription

Smart constructor for the minimum required fields to construct a valid IdentityPoolShortDescription data type.

IdentityPoolShortDescription is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IdentityPoolId :: Maybe Text

IdentityPoolName :: Maybe Text

ipsdIdentityPoolId :: Lens' IdentityPoolShortDescription (Maybe Text)

An identity pool ID in the format REGION:GUID.

ipsdIdentityPoolName :: Lens' IdentityPoolShortDescription (Maybe Text)

A string that you provide.