Network.AWS.CognitoIdentity.V2014_06_30.UpdateIdentityPool

Description

Updates a user pool. UpdateIdentityPool The following are a request and response for the UpdateIdentityPool action. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyUpdatedPool, IdentityPoolDescription: "An identity pool that needs updating", Unauthenticated: true, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID } } { IdentityPoolDescription: "An identity pool that needs updating", IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyUpdatedPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, AllowUnauthenticatedIdentities: true }.

Documentation

updateIdentityPool

Arguments

:: Text
_irIdentityPoolId

-> Text
_irIdentityPoolName

-> Bool
_irAllowUnauthenticatedIdentities

-> UpdateIdentityPool

Minimum specification for a UpdateIdentityPool request.

data UpdateIdentityPool

Constructors

UpdateIdentityPool

Fields

_irIdentityPoolId :: Text

An identity pool ID in the format REGION:GUID.

_irIdentityPoolName :: Text

A string that you provide.

_irAllowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

_irSupportedLoginProviders :: Map Text Text

Optional key:value pairs mapping provider names to provider app IDs.

irSupportedLoginProviders :: Lens' UpdateIdentityPool (Map Text Text)

irIdentityPoolName :: Lens' UpdateIdentityPool Text

irIdentityPoolId :: Lens' UpdateIdentityPool Text

irAllowUnauthenticatedIdentities :: Lens' UpdateIdentityPool Bool

data UpdateIdentityPoolResponse

Constructors

UpdateIdentityPoolResponse

Fields

_itIdentityPoolId :: Text

An identity pool ID in the format REGION:GUID.

_itIdentityPoolName :: Text

A string that you provide.

_itAllowUnauthenticatedIdentities :: Bool

TRUE if the identity pool supports unauthenticated logins.

_itSupportedLoginProviders :: Map Text Text

Optional key:value pairs mapping provider names to provider app IDs.

itSupportedLoginProviders :: Lens' UpdateIdentityPoolResponse (Map Text Text)

itIdentityPoolName :: Lens' UpdateIdentityPoolResponse Text

itIdentityPoolId :: Lens' UpdateIdentityPoolResponse Text

itAllowUnauthenticatedIdentities :: Lens' UpdateIdentityPoolResponse Bool