Network.AWS.CognitoIdentity.DescribeIdentity

Description

Returns metadata related to the given identity, including when the identity was created and any associated linked logins.

You must use AWS Developer credentials to call this API.

Creating a Request

describeIdentity

Arguments

:: Text
diIdentityId

-> DescribeIdentity

Creates a value of DescribeIdentity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diIdentityId - A unique identifier in the format REGION:GUID.

data DescribeIdentity

Input to the DescribeIdentity action.

See: describeIdentity smart constructor.

Request Lenses

diIdentityId :: Lens' DescribeIdentity Text

A unique identifier in the format REGION:GUID.

Destructuring the Response

identityDescription :: IdentityDescription

Creates a value of IdentityDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

idLastModifiedDate - Date on which the identity was last modified.

idCreationDate - Date on which the identity was created.

idLogins - A set of optional name-value pairs that map provider names to provider tokens.

idIdentityId - A unique identifier in the format REGION:GUID.

data IdentityDescription

A description of the identity.

See: identityDescription smart constructor.

Response Lenses

idLastModifiedDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was last modified.

idCreationDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was created.

idLogins :: Lens' IdentityDescription [Text]

A set of optional name-value pairs that map provider names to provider tokens.

idIdentityId :: Lens' IdentityDescription (Maybe Text)

A unique identifier in the format REGION:GUID.