Network.AWS.CognitoIdentity.DescribeIdentityPool

Description

Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users.

You must use AWS Developer credentials to call this API.

Creating a Request

describeIdentityPool

Arguments

:: Text
dipIdentityPoolId

-> DescribeIdentityPool

Creates a value of DescribeIdentityPool with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dipIdentityPoolId - An identity pool ID in the format REGION:GUID.

data DescribeIdentityPool

Input to the DescribeIdentityPool action.

See: describeIdentityPool smart constructor.

Request Lenses

dipIdentityPoolId :: Lens' DescribeIdentityPool Text

An identity pool ID in the format REGION:GUID.

Destructuring the Response

identityPool

Arguments

:: Text
ipIdentityPoolId

-> Text
ipIdentityPoolName

-> Bool
ipAllowUnauthenticatedIdentities

-> IdentityPool

Creates a value of IdentityPool with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ipSamlProviderARNs - An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

ipSupportedLoginProviders - Optional key:value pairs mapping provider names to provider app IDs.

ipDeveloperProviderName - The "domain" by which Cognito will refer to your users.

ipOpenIdConnectProviderARNs - A list of OpendID Connect provider ARNs.

ipCognitoIdentityProviders - A list representing an Amazon Cognito Identity User Pool and its client ID.

ipIdentityPoolId - An identity pool ID in the format REGION:GUID.

ipIdentityPoolName - A string that you provide.

ipAllowUnauthenticatedIdentities - TRUE if the identity pool supports unauthenticated logins.

data IdentityPool

An object representing an Amazon Cognito identity pool.

See: identityPool smart constructor.

Response Lenses

ipSamlProviderARNs :: Lens' IdentityPool [Text]

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

ipSupportedLoginProviders :: Lens' IdentityPool (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

ipDeveloperProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

ipOpenIdConnectProviderARNs :: Lens' IdentityPool [Text]

A list of OpendID Connect provider ARNs.

ipCognitoIdentityProviders :: Lens' IdentityPool [CognitoIdentityProvider]

A list representing an Amazon Cognito Identity User Pool and its client ID.

ipIdentityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

ipIdentityPoolName :: Lens' IdentityPool Text

A string that you provide.

ipAllowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.