Network.AWS.CognitoIdentity.DescribeIdentityPool

Description

Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_DescribeIdentityPool.html

Request

data DescribeIdentityPool

Request constructor

describeIdentityPool

Arguments

:: Text
dipIdentityPoolId

-> DescribeIdentityPool

DescribeIdentityPool constructor.

The fields accessible through corresponding lenses are:

dipIdentityPoolId :: Text

Request lenses

dipIdentityPoolId :: Lens' DescribeIdentityPool Text

An identity pool ID in the format REGION:GUID.

Response

data DescribeIdentityPoolResponse

Response constructor

describeIdentityPoolResponse

Arguments

:: Text
diprIdentityPoolId

-> Text
diprIdentityPoolName

-> Bool
diprAllowUnauthenticatedIdentities

-> DescribeIdentityPoolResponse

DescribeIdentityPoolResponse constructor.

The fields accessible through corresponding lenses are:

diprAllowUnauthenticatedIdentities :: Bool

diprDeveloperProviderName :: Maybe Text

diprIdentityPoolId :: Text

diprIdentityPoolName :: Text

diprOpenIdConnectProviderARNs :: [Text]

diprSupportedLoginProviders :: HashMap Text Text

Response lenses

diprAllowUnauthenticatedIdentities :: Lens' DescribeIdentityPoolResponse Bool

TRUE if the identity pool supports unauthenticated logins.

diprDeveloperProviderName :: Lens' DescribeIdentityPoolResponse (Maybe Text)

The "domain" by which Cognito will refer to your users.

diprIdentityPoolId :: Lens' DescribeIdentityPoolResponse Text

An identity pool ID in the format REGION:GUID.

diprIdentityPoolName :: Lens' DescribeIdentityPoolResponse Text

A string that you provide.

diprOpenIdConnectProviderARNs :: Lens' DescribeIdentityPoolResponse [Text]

A list of OpendID Connect provider ARNs.

diprSupportedLoginProviders :: Lens' DescribeIdentityPoolResponse (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.