Network.AWS.CognitoIdentity.GetId

Description

Generates (or retrieves) a Cognito ID. Supplying multiple logins will create an implicit linked account.

This is a public API. You do not need any credentials to call this API.

Creating a Request

getId

Arguments

:: Text
giIdentityPoolId

-> GetId

Creates a value of GetId with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giAccountId - A standard AWS account ID (9+ digits).

giLogins - A set of optional name-value pairs that map provider names to provider tokens. The available provider names for Logins are as follows: * Facebook: graph.facebook.com * Amazon Cognito Identity Provider: cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789 * Google: accounts.google.com * Amazon: www.amazon.com * Twitter: api.twitter.com * Digits: www.digits.com

giIdentityPoolId - An identity pool ID in the format REGION:GUID.

data GetId

Input to the GetId action.

See: getId smart constructor.

Request Lenses

giAccountId :: Lens' GetId (Maybe Text)

A standard AWS account ID (9+ digits).

giLogins :: Lens' GetId (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens. The available provider names for Logins are as follows: * Facebook: graph.facebook.com * Amazon Cognito Identity Provider: cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789 * Google: accounts.google.com * Amazon: www.amazon.com * Twitter: api.twitter.com * Digits: www.digits.com

giIdentityPoolId :: Lens' GetId Text

An identity pool ID in the format REGION:GUID.

Destructuring the Response

getIdResponse

Arguments

:: Int
girsResponseStatus

-> GetIdResponse

Creates a value of GetIdResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

girsIdentityId - A unique identifier in the format REGION:GUID.

girsResponseStatus - -- | The response status code.

data GetIdResponse

Returned in response to a GetId request.

See: getIdResponse smart constructor.

Response Lenses

girsIdentityId :: Lens' GetIdResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

girsResponseStatus :: Lens' GetIdResponse Int

- | The response status code.