Network.AWS.CognitoIdentity.GetOpenIdToken

Description

Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned by GetId . You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link.

The OpenId token is valid for 15 minutes.

This is a public API. You do not need any credentials to call this API.

Creating a Request

getOpenIdToken

Arguments

:: Text
goitIdentityId

-> GetOpenIdToken

Creates a value of GetOpenIdToken with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

goitLogins - A set of optional name-value pairs that map provider names to provider tokens. When using graph.facebook.com and www.amazon.com, supply the access_token returned from the provider's authflow. For accounts.google.com, an Amazon Cognito Identity Provider, or any other OpenId Connect provider, always include the id_token .

goitIdentityId - A unique identifier in the format REGION:GUID.

data GetOpenIdToken

Input to the GetOpenIdToken action.

See: getOpenIdToken smart constructor.

Request Lenses

goitLogins :: Lens' GetOpenIdToken (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens. When using graph.facebook.com and www.amazon.com, supply the access_token returned from the provider's authflow. For accounts.google.com, an Amazon Cognito Identity Provider, or any other OpenId Connect provider, always include the id_token .

goitIdentityId :: Lens' GetOpenIdToken Text

A unique identifier in the format REGION:GUID.

Destructuring the Response

getOpenIdTokenResponse

Arguments

:: Int
goitrsResponseStatus

-> GetOpenIdTokenResponse

Creates a value of GetOpenIdTokenResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

goitrsToken - An OpenID token, valid for 15 minutes.

goitrsIdentityId - A unique identifier in the format REGION:GUID. Note that the IdentityId returned may not match the one passed on input.

goitrsResponseStatus - -- | The response status code.

data GetOpenIdTokenResponse

Returned in response to a successful GetOpenIdToken request.

See: getOpenIdTokenResponse smart constructor.

Response Lenses

goitrsToken :: Lens' GetOpenIdTokenResponse (Maybe Text)

An OpenID token, valid for 15 minutes.

goitrsIdentityId :: Lens' GetOpenIdTokenResponse (Maybe Text)

A unique identifier in the format REGION:GUID. Note that the IdentityId returned may not match the one passed on input.

goitrsResponseStatus :: Lens' GetOpenIdTokenResponse Int

- | The response status code.