Network.AWS.CognitoIdentity.GetOpenIdToken

Description

Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned by GetId. You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link.

The OpenId token is valid for 15 minutes.

http://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_GetOpenIdToken.html

Request

data GetOpenIdToken

Request constructor

getOpenIdToken

Arguments

:: Text
goitIdentityId

-> GetOpenIdToken

GetOpenIdToken constructor.

The fields accessible through corresponding lenses are:

goitIdentityId :: Text

goitLogins :: HashMap Text Text

Request lenses

goitIdentityId :: Lens' GetOpenIdToken Text

A unique identifier in the format REGION:GUID.

goitLogins :: Lens' GetOpenIdToken (HashMap Text Text)

A set of optional name-value pairs that map provider names to provider tokens.

Response

data GetOpenIdTokenResponse

Response constructor

getOpenIdTokenResponse :: GetOpenIdTokenResponse

GetOpenIdTokenResponse constructor.

The fields accessible through corresponding lenses are:

goitrIdentityId :: Maybe Text

goitrToken :: Maybe Text

Response lenses

goitrIdentityId :: Lens' GetOpenIdTokenResponse (Maybe Text)

A unique identifier in the format REGION:GUID. Note that the IdentityId returned may not match the one passed on input.

goitrToken :: Lens' GetOpenIdTokenResponse (Maybe Text)

An OpenID token, valid for 15 minutes.