Network.AWS.CognitoIdentity.LookupDeveloperIdentity

Description

Retrieves the IdentityID associated with a DeveloperUserIdentifier or the list of DeveloperUserIdentifier s associated with an IdentityId for an existing identity. Either IdentityID or DeveloperUserIdentifier must not be null. If you supply only one of these values, the other value will be searched in the database and returned as a part of the response. If you supply both, DeveloperUserIdentifier will be matched against IdentityID . If the values are verified against the database, the response returns both values and is the same as the request. Otherwise a ResourceConflictException is thrown.

You must use AWS Developer credentials to call this API.

Creating a Request

lookupDeveloperIdentity

Arguments

:: Text
ldiIdentityPoolId

-> LookupDeveloperIdentity

Creates a value of LookupDeveloperIdentity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldiDeveloperUserIdentifier - A unique ID used by your backend authentication process to identify a user. Typically, a developer identity provider would issue many developer user identifiers, in keeping with the number of users.

ldiNextToken - A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

ldiIdentityId - A unique identifier in the format REGION:GUID.

ldiMaxResults - The maximum number of identities to return.

ldiIdentityPoolId - An identity pool ID in the format REGION:GUID.

data LookupDeveloperIdentity

Input to the LookupDeveloperIdentityInput action.

See: lookupDeveloperIdentity smart constructor.

Request Lenses

ldiDeveloperUserIdentifier :: Lens' LookupDeveloperIdentity (Maybe Text)

A unique ID used by your backend authentication process to identify a user. Typically, a developer identity provider would issue many developer user identifiers, in keeping with the number of users.

ldiNextToken :: Lens' LookupDeveloperIdentity (Maybe Text)

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

ldiIdentityId :: Lens' LookupDeveloperIdentity (Maybe Text)

A unique identifier in the format REGION:GUID.

ldiMaxResults :: Lens' LookupDeveloperIdentity (Maybe Natural)

The maximum number of identities to return.

ldiIdentityPoolId :: Lens' LookupDeveloperIdentity Text

An identity pool ID in the format REGION:GUID.

Destructuring the Response

lookupDeveloperIdentityResponse

Arguments

:: Int
ldirsResponseStatus

-> LookupDeveloperIdentityResponse

Creates a value of LookupDeveloperIdentityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldirsNextToken - A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

ldirsIdentityId - A unique identifier in the format REGION:GUID.

ldirsDeveloperUserIdentifierList - This is the list of developer user identifiers associated with an identity ID. Cognito supports the association of multiple developer user identifiers with an identity ID.

ldirsResponseStatus - -- | The response status code.

data LookupDeveloperIdentityResponse

Returned in response to a successful LookupDeveloperIdentity action.

See: lookupDeveloperIdentityResponse smart constructor.

Response Lenses

ldirsNextToken :: Lens' LookupDeveloperIdentityResponse (Maybe Text)

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

ldirsIdentityId :: Lens' LookupDeveloperIdentityResponse (Maybe Text)

A unique identifier in the format REGION:GUID.

ldirsDeveloperUserIdentifierList :: Lens' LookupDeveloperIdentityResponse [Text]

This is the list of developer user identifiers associated with an identity ID. Cognito supports the association of multiple developer user identifiers with an identity ID.

ldirsResponseStatus :: Lens' LookupDeveloperIdentityResponse Int

- | The response status code.