Network.AWS.CognitoIdentity.Types

Service

data CognitoIdentity

Version 2014-06-30 of the Amazon Cognito Identity service.

Error

data JSONError :: *

IdentityDescription

data IdentityDescription

identityDescription :: IdentityDescription

IdentityDescription constructor.

The fields accessible through corresponding lenses are:

idIdentityId :: Maybe Text

idLogins :: [Text]

idIdentityId :: Lens' IdentityDescription (Maybe Text)

A unique identifier in the format REGION:GUID.

idLogins :: Lens' IdentityDescription [Text]

A set of optional name-value pairs that map provider names to provider tokens.

IdentityPool

data IdentityPool

identityPool

Arguments

:: Text
ipIdentityPoolId

-> Text
ipIdentityPoolName

-> Bool
ipAllowUnauthenticatedIdentities

-> IdentityPool

IdentityPool constructor.

The fields accessible through corresponding lenses are:

ipAllowUnauthenticatedIdentities :: Bool

ipDeveloperProviderName :: Maybe Text

ipIdentityPoolId :: Text

ipIdentityPoolName :: Text

ipOpenIdConnectProviderARNs :: [Text]

ipSupportedLoginProviders :: HashMap Text Text

ipAllowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

ipDeveloperProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

ipIdentityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

ipIdentityPoolName :: Lens' IdentityPool Text

A string that you provide.

ipOpenIdConnectProviderARNs :: Lens' IdentityPool [Text]

ipSupportedLoginProviders :: Lens' IdentityPool (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

IdentityPoolShortDescription

data IdentityPoolShortDescription

identityPoolShortDescription :: IdentityPoolShortDescription

IdentityPoolShortDescription constructor.

The fields accessible through corresponding lenses are:

ipsdIdentityPoolId :: Maybe Text

ipsdIdentityPoolName :: Maybe Text

ipsdIdentityPoolId :: Lens' IdentityPoolShortDescription (Maybe Text)

An identity pool ID in the format REGION:GUID.

ipsdIdentityPoolName :: Lens' IdentityPoolShortDescription (Maybe Text)

A string that you provide.