Network.AWS.CognitoIdentity.Types

Service

data CognitoIdentity

Version 2014-06-30 of the Amazon Cognito Identity service.

Error

data JSONError :: *

Credentials

data Credentials

credentials :: Credentials

Credentials constructor.

The fields accessible through corresponding lenses are:

cAccessKeyId :: Maybe Text

cExpiration :: Maybe UTCTime

cSecretKey :: Maybe Text

cSessionToken :: Maybe Text

cAccessKeyId :: Lens' Credentials (Maybe Text)

The Access Key portion of the credentials.

cExpiration :: Lens' Credentials (Maybe UTCTime)

The date at which these credentials will expire.

cSecretKey :: Lens' Credentials (Maybe Text)

The Secret Access Key portion of the credentials

cSessionToken :: Lens' Credentials (Maybe Text)

The Session Token portion of the credentials

IdentityDescription

data IdentityDescription

identityDescription :: IdentityDescription

IdentityDescription constructor.

The fields accessible through corresponding lenses are:

idCreationDate :: Maybe UTCTime

idIdentityId :: Maybe Text

idLastModifiedDate :: Maybe UTCTime

idLogins :: [Text]

idCreationDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was created.

idIdentityId :: Lens' IdentityDescription (Maybe Text)

A unique identifier in the format REGION:GUID.

idLastModifiedDate :: Lens' IdentityDescription (Maybe UTCTime)

Date on which the identity was last modified.

idLogins :: Lens' IdentityDescription [Text]

A set of optional name-value pairs that map provider names to provider tokens.

IdentityPool

data IdentityPool

identityPool

Arguments

:: Text
ipIdentityPoolId

-> Text
ipIdentityPoolName

-> Bool
ipAllowUnauthenticatedIdentities

-> IdentityPool

IdentityPool constructor.

The fields accessible through corresponding lenses are:

ipAllowUnauthenticatedIdentities :: Bool

ipDeveloperProviderName :: Maybe Text

ipIdentityPoolId :: Text

ipIdentityPoolName :: Text

ipOpenIdConnectProviderARNs :: [Text]

ipSupportedLoginProviders :: HashMap Text Text

ipAllowUnauthenticatedIdentities :: Lens' IdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

ipDeveloperProviderName :: Lens' IdentityPool (Maybe Text)

The "domain" by which Cognito will refer to your users.

ipIdentityPoolId :: Lens' IdentityPool Text

An identity pool ID in the format REGION:GUID.

ipIdentityPoolName :: Lens' IdentityPool Text

A string that you provide.

ipOpenIdConnectProviderARNs :: Lens' IdentityPool [Text]

A list of OpendID Connect provider ARNs.

ipSupportedLoginProviders :: Lens' IdentityPool (HashMap Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

IdentityPoolShortDescription

data IdentityPoolShortDescription

identityPoolShortDescription :: IdentityPoolShortDescription

IdentityPoolShortDescription constructor.

The fields accessible through corresponding lenses are:

ipsdIdentityPoolId :: Maybe Text

ipsdIdentityPoolName :: Maybe Text

ipsdIdentityPoolId :: Lens' IdentityPoolShortDescription (Maybe Text)

An identity pool ID in the format REGION:GUID.

ipsdIdentityPoolName :: Lens' IdentityPoolShortDescription (Maybe Text)

A string that you provide.