Network.AWS.CognitoIdentity.V2014_06_30.CreateIdentityPool

Description

Creates a new identity pool. The identity pool is a store of user identity information that is specific to your AWS account. CreateIdentityPool The following example shows a request and response for a CreateIdentityPool operation. { IdentityPoolName: MyIdentityPool, IdentityPoolDescription: "My identity pool", Unauthenticated: true, SupportedLoginProviders: { "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID, "www.amazon.com": Amazon_App_ID } } { IdentityPoolDescription: "My identity pool", IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyIdentityPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, Unauthenticated: true }.

Request

data CreateIdentityPool

Input to the CreateIdentityPool action.

Request constructor

mkCreateIdentityPool

Arguments

:: Text
cipIdentityPoolName

-> Bool
cipAllowUnauthenticatedIdentities

-> CreateIdentityPool

Smart constructor for the minimum required parameters to construct a valid CreateIdentityPool request.

Request lenses

cipIdentityPoolName :: Lens' CreateIdentityPool Text

A string that you provide.

cipAllowUnauthenticatedIdentities :: Lens' CreateIdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

cipSupportedLoginProviders :: Lens' CreateIdentityPool (Map Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

Response

data CreateIdentityPoolResponse

An object representing a Cognito identity pool.

Response lenses

ciprsIdentityPoolId :: Lens' CreateIdentityPoolResponse Text

An identity pool ID in the format REGION:GUID.

ciprsIdentityPoolName :: Lens' CreateIdentityPoolResponse Text

A string that you provide.

ciprsAllowUnauthenticatedIdentities :: Lens' CreateIdentityPoolResponse Bool

TRUE if the identity pool supports unauthenticated logins.

ciprsSupportedLoginProviders :: Lens' CreateIdentityPoolResponse (Map Text Text)

Optional key:value pairs mapping provider names to provider app IDs.