Network.AWS.CognitoIdentity.V2014_06_30.DeleteIdentityPool

Description

Deletes a user pool. Once a pool is deleted, users will not be able to authenticate with the pool. DeleteIdentityPool The following is an example of a DeleteIdentityPool request. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE" }.

Request

data DeleteIdentityPool

Input to the DeleteIdentityPool action.

Request constructor

mkDeleteIdentityPool

Arguments

:: Text
dipIdentityPoolId

-> DeleteIdentityPool

Smart constructor for the minimum required parameters to construct a valid DeleteIdentityPool request.

The fields accessible through corresponding lenses are:

IdentityPoolId :: Text

Request lenses

dipIdentityPoolId :: Lens' DeleteIdentityPool Text

An identity pool ID in the format REGION:GUID.

Response

data DeleteIdentityPoolResponse

Response constructor

mkDeleteIdentityPoolResponse :: DeleteIdentityPoolResponse

Smart constructor for the minimum required parameters to construct a valid DeleteIdentityPoolResponse response.

This constructor is provided for convenience and testing purposes.