Network.AWS.CognitoIdentity.V2014_06_30.DescribeIdentityPool

Description

Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users. DescribeIdentityPool The following are an example request and response for the DescribeIdentityPool operation. { IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1" } { IdentityPoolDescription: "My identity pool", IdentityPoolId: "us-east-1:af4311ca-835e-4b49-814c-2290EXAMPLE1", IdentityPoolName: MyIdentityPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, Unauthenticated: true }.

Request

data DescribeIdentityPool

Input to the DescribeIdentityPool action.

Request constructor

mkDescribeIdentityPool

Arguments

:: Text
dip1IdentityPoolId

-> DescribeIdentityPool

Smart constructor for the minimum required parameters to construct a valid DescribeIdentityPool request.

The fields accessible through corresponding lenses are:

IdentityPoolId :: Text

Request lenses

dip1IdentityPoolId :: Lens' DescribeIdentityPool Text

An identity pool ID in the format REGION:GUID.

Response

data DescribeIdentityPoolResponse

An object representing a Cognito identity pool.

Response constructor

mkDescribeIdentityPoolResponse

Arguments

:: Text
diprIdentityPoolId

-> Text
diprIdentityPoolName

-> Bool
diprAllowUnauthenticatedIdentities

-> DescribeIdentityPoolResponse

Smart constructor for the minimum required parameters to construct a valid DescribeIdentityPoolResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IdentityPoolId :: Text

IdentityPoolName :: Text

AllowUnauthenticatedIdentities :: Bool

SupportedLoginProviders :: Map Text Text

Response lenses

diprIdentityPoolId :: Lens' DescribeIdentityPoolResponse Text

An identity pool ID in the format REGION:GUID.

diprIdentityPoolName :: Lens' DescribeIdentityPoolResponse Text

A string that you provide.

diprAllowUnauthenticatedIdentities :: Lens' DescribeIdentityPoolResponse Bool

TRUE if the identity pool supports unauthenticated logins.

diprSupportedLoginProviders :: Lens' DescribeIdentityPoolResponse (Map Text Text)

Optional key:value pairs mapping provider names to provider app IDs.