Network.AWS.CognitoIdentity.V2014_06_30.ListIdentities

Description

Lists the identities in a pool. ListIdentities The following are examples of a request and a response for the ListIdentities action. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", MaxResults: 10 } { Identities: [ { IdentityId: "us-east-1:2345a6b7-8cc3-4a60-8aeb-e11bEXAMPLE4" }, { IdentityId: "us-east-1:852d4250-9eec-4006-8f84-4e82EXAMPLE3" }, { IdentityId: "us-east-1:921a3843-2dd6-46b8-ab2f-c679EXAMPLE5" } ], IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE" }.

Request

data ListIdentities

Input to the ListIdentities action.

Request constructor

mkListIdentities

Arguments

:: Text
liIdentityPoolId

-> Integer
liMaxResults

-> ListIdentities

Smart constructor for the minimum required parameters to construct a valid ListIdentities request.

Request lenses

liIdentityPoolId :: Lens' ListIdentities Text

An identity pool ID in the format REGION:GUID.

liMaxResults :: Lens' ListIdentities Integer

The maximum number of identities to return.

liNextToken :: Lens' ListIdentities (Maybe Text)

A pagination token.

Response

data ListIdentitiesResponse

The response to a ListIdentities request.

Response lenses

lirsIdentityPoolId :: Lens' ListIdentitiesResponse (Maybe Text)

An identity pool ID in the format REGION:GUID.

lirsIdentities :: Lens' ListIdentitiesResponse [IdentityDescription]

An object containing a set of identities and associated mappings.

lirsNextToken :: Lens' ListIdentitiesResponse (Maybe Text)

A pagination token.