Network.AWS.CognitoIdentity.V2014_06_30.UpdateIdentityPool

Description

Updates a user pool. UpdateIdentityPool The following are a request and response for the UpdateIdentityPool action. { IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyUpdatedPool, IdentityPoolDescription: "An identity pool that needs updating", Unauthenticated: true, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID } } { IdentityPoolDescription: "An identity pool that needs updating", IdentityPoolId: "us-east-1:1a234b56-7890-1cd2-3e45-f6g7hEXAMPLE", IdentityPoolName: MyUpdatedPool, SupportedLoginProviders: { "www.amazon.com": Amazon_App_ID, "graph.facebook.com": Facebook_App_ID, "accounts.google.com": Google_App_ID }, AllowUnauthenticatedIdentities: true }.

Request

data UpdateIdentityPool

An object representing a Cognito identity pool.

Request constructor

mkUpdateIdentityPool

Arguments

:: Text
uipIdentityPoolId

-> Text
uipIdentityPoolName

-> Bool
uipAllowUnauthenticatedIdentities

-> UpdateIdentityPool

Smart constructor for the minimum required parameters to construct a valid UpdateIdentityPool request.

Request lenses

uipIdentityPoolId :: Lens' UpdateIdentityPool Text

An identity pool ID in the format REGION:GUID.

uipIdentityPoolName :: Lens' UpdateIdentityPool Text

A string that you provide.

uipAllowUnauthenticatedIdentities :: Lens' UpdateIdentityPool Bool

TRUE if the identity pool supports unauthenticated logins.

uipSupportedLoginProviders :: Lens' UpdateIdentityPool (Map Text Text)

Optional key:value pairs mapping provider names to provider app IDs.

Response

data UpdateIdentityPoolResponse

An object representing a Cognito identity pool.

Response lenses

uiprIdentityPoolId :: Lens' UpdateIdentityPoolResponse Text

An identity pool ID in the format REGION:GUID.

uiprIdentityPoolName :: Lens' UpdateIdentityPoolResponse Text

A string that you provide.

uiprAllowUnauthenticatedIdentities :: Lens' UpdateIdentityPoolResponse Bool

TRUE if the identity pool supports unauthenticated logins.

uiprSupportedLoginProviders :: Lens' UpdateIdentityPoolResponse (Map Text Text)

Optional key:value pairs mapping provider names to provider app IDs.